backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pydantic==2.10.3
email-validator==2.2.0
httpx==0.28.1
PyMuPDF==1.25.1
python-docx==1.1.2
pytest==8.3.4
